Only download translation strings we have defined
